+++ openstack/blockstorage/v1/volumetypes/delete.go
@@ -1,0 +1,8 @@
+package volumetypes
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("types", id), nil)
+	return
+}
