+++ openstack/blockstorage/v1/volumetypes/testing/requests_test.go
@@ -58,7 +58,7 @@ func TestGet(t *testing.T) {
 
 	MockGetResponse(t)
 
-	vt, err := volumetypes.Get(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22").Extract()
+	vt, err := volumetypes.Get(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22")
 	th.AssertNoErr(t, err)
 
 	th.AssertDeepEquals(t, vt.ExtraSpecs, map[string]interface{}{"serverNumber": "2"})
@@ -96,8 +96,8 @@ func TestCreate(t *testing.T) {
 		`)
 	})
 
-	options := &volumetypes.CreateOpts{Name: "vol-type-001"}
-	n, err := volumetypes.Create(client.ServiceClient(), options).Extract()
+	options := volumetypes.CreateOpts{Name: "vol-type-001"}
+	n, err := volumetypes.Create(client.ServiceClient(), options)
 	th.AssertNoErr(t, err)
 
 	th.AssertEquals(t, n.Name, "vol-type-001")
@@ -114,6 +114,6 @@ func TestDelete(t *testing.T) {
 		w.WriteHeader(http.StatusAccepted)
 	})
 
-	err := volumetypes.Delete(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22").ExtractErr()
+	err := volumetypes.Delete(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22")
 	th.AssertNoErr(t, err)
 }
